Move derived classes to their own files
